# docs/swagger.yaml
basePath: /api/v1
definitions:
  Tokens:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  User:
    properties:
      guid:
        type: string
    type: object
  models.RefreshToken:
    properties:
      refresh_token:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a simple API project
  title: Test task BackDev
  version: "1.0"
paths:
  /authenticate:
    get:
      consumes:
      - application/json
      description: Generate pair of tokens
      parameters:
      - description: User guid
        in: query
        name: guid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Tokens'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Generate pair of tokens
      tags:
      - auth
  /refresh:
    post:
      consumes:
      - application/json
      description: Refresh token
      parameters:
      - description: Refresh token
        in: body
        name: refresh_token
        required: true
        schema:
          $ref: '#/definitions/models.RefreshToken'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Tokens'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Refresh token
      tags:
      - auth
  /register:
    post:
      consumes:
      - application/json
      description: Register new user
      parameters:
      - description: User object to be added
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/User'
      produces:
      - application/json
      responses:
        "201":
          description: User created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Register new user
      tags:
      - auth
swagger: "2.0"